Started dev on scrum activity
