+++ client/src/components/exercise/lab0/DesignLabSection/DesignLabIntroduction.js
@@ -1,7 +1,13 @@
 import React from "react";
 
 const DesignLabIntroduction = () => {
-  return <div>Design Lab</div>;
+  return (
+    <div>
+      Design LabDesign LabDesign LabDesign LabDesign LabDesign LabDesign
+      LabDesign LabDesign LabDesign LabDesign LabDesign LabDesign LabDesign
+      LabDesign Lab
+    </div>
+  );
 };
 
 export default DesignLabIntroduction;
