+++ package.json
@@ -26,6 +26,7 @@
     "react-helmet": "^6.1.0",
     "react-highlight": "^0.15.0",
     "react-live": "^4.0.0",
+    "react-router-dom": "^7.1.5",
     "react-syntax-highlighter": "^15.5.0",
     "reactstrap": "^9.2.2",
     "redux-logger": "^3.0.6",
